+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2.0.8-rc (release candidate)
+# 2.0.8-rc (2024-08-08 20:23)
 
 ### Features